Refactored SaveModelWekaBatchTask to externalize the actual task for storing a model (so that this task can be used elsewhere).
